pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-filter>=25.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "whitenoise>=6.11.0",
]

[build-system] 
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel] 
packages = ["task_manager"]
